First Commit, adding all
Music player working with 2 examples - 1 is the star wars theme, the
other is a single note being played.

Known issues
* after playing the single note, the buzzer turns off then starts to
put out static sounds. Not sure what’s causing it.
* if you’re playing music while using RGB LED, it causes interference
with the music & causes the LED color to be altered. Still
investigating, but seems to be related to PWM issue mentioned here -
https://www.arduino.cc/en/Reference/Tone